Move the logic to Internal module and expose everything inside.
The resource-pool package didn't expose its internals. It leads
to a situation when the libraries has no required introspection
into the package and can't implement additional functionality without
changing the resource-pool itself. Examples of such functionality:

   1. minimal amount of allocated resources
   2. setting a cap on the number of concurrently opened resources
   3. adding metrics

All those features can fit the package nicely and should be implented
but having internal exposed provide cheap way to extend the package
without waiting for a new version to come, because there may be more
features that is too hard to imagine beforehand.